#185 [REFACTOR] : RE-packaging charo - old version, mypage - new version
