+++ app/src/main/java/com/example/charo_android/data/api/mypage/MyPageViewFollowService.kt
@@ -1,6 +1,6 @@
 package com.example.charo_android.data.api.mypage
 
-import com.example.charo_android.data.model.mypage.ResponseMyPageFollowData
+import com.example.charo_android.data.model.charo.ResponseMyPageFollowData
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
