+++ app/src/main/java/com/example/charo_android/data/api/mypage/MyPageViewMoreService.kt
@@ -1,6 +1,6 @@
 package com.example.charo_android.data.api.mypage
 
-import com.example.charo_android.data.model.mypage.ResponseMyPageMoreData
+import com.example.charo_android.data.model.charo.ResponseMyPageMoreData
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
