+++ app/src/main/java/com/example/charo_android/data/model/charo/Drive.kt
@@ -1,5 +1,5 @@
 
-package com.example.charo_android.data.model.mypage
+package com.example.charo_android.data.model.charo
 
 data class Drive(
     val day: String,
