+++ app/src/main/java/com/example/charo_android/data/model/charo/ResponseMyPageLikeData.kt
@@ -1,4 +1,4 @@
-package com.example.charo_android.data.model.mypage
+package com.example.charo_android.data.model.charo
 
 data class ResponseMyPageLikeData(
     val data: Data,
