+++ app/src/main/java/com/example/charo_android/data/model/mypage/SampleDataClass.kt
@@ -1,0 +1,5 @@
+package com.example.charo_android.data.model.mypage
+
+data class SampleDataClass(
+    val id: Int,
+)
