--- app/src/main/java/com/example/charo_android/data/model/newcharo/SampleDataClass.kt
@@ -1,5 +1,0 @@
-package com.example.charo_android.data.model.newcharo
-
-data class SampleDataClass(
-    val id: Int,
-)
