+++ app/src/main/res/layout/fragment_profile.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.newcharo.ProfileFragment">
+        tools:context=".presentation.ui.mypage.ProfileFragment">
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fcv_top"
