+++ app/src/main/res/layout/fragment_written_post.xml
@@ -10,7 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.ui.newcharo.list.WrittenPostFragment">
+        tools:context=".presentation.ui.mypage.list.WrittenPostFragment">
 
         <androidx.core.widget.NestedScrollView
             android:id="@+id/nsv_post_list"
